% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_disease_scores_all_time_v2.R
\name{calculate_disease_scores}
\alias{calculate_disease_scores}
\title{calculate_disease_scores}
\usage{
calculate_disease_scores(datadir = ".")
}
\arguments{
\item{datadir}{directory where unzipped data is saved.}
}
\value{
A dataframe with the scores as different tabs and a list of dataframes with each score as a list element.
}
\description{
Calculates Latest Diagnosis, sCDAI, 6 POINT MAYO, Manitoba, & PGA for SPARC cohort
}
