% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_calculations.R
\name{calculate_location_az}
\alias{calculate_location_az}
\title{calculate_location_az}
\usage{
calculate_location_az(observations, encounter, window = 90)
}
\arguments{
\item{observations}{observations table usually generated using load_data}

\item{encounter}{encounter table usually generated using load_data}

\item{window}{Time window (in days) in which observations labeled "unknown" can be carried over.}
}
\value{
A dataframe with all CD locations that could be calculated regardless of IBD diagnosis.
}
\description{
Calculates disease locations from SPARC data with criteria specific to AZ
}
\details{
Contradictions and ambiguities in a location's involvement are resolved as follows:
If more than one record for a location exists for the same patient on the
same date, any record of a "yes" is taken as evidence for this location
showing disease activity. In absence of any "yes", any record of a "no" is
taken of evidence of absence of evidence. Otherwise, any occurrence of
"unknown" among the records is collapsed to a single "unknown". Any remaining
combination of values is considered "INVALID".

The window variable can be used to leverage the relative stability of the
involvement of locations in Crohn's disease. Unknown values
will be replaced if other observations exist in the specified time window.
Only Observations in the future of the visit with missing records will be
considered.

Anal involvement is considered as contributing to a colonic phenotype and not as a perianal modifier.

In addition of the recorded CD locations in SF observations, we can use records of strictures to
add certainty to "unknown" disease activities or even correct records of absence of disease activity.
Any record of strictures within seven days of the date of the record of the CD location will be considered.
\itemize{
 \item{If there is a record of "Phenotype - Ileal Stricture" = "Yes"
 it's considered as evidence of disease activity in the Ileum.}
 \item{Any records of "Phenotype - Left Colonic Stricture",
 "Phenotype - Rectal Stricture", "Phenotype - Right Colonic Stricture" or
 "Phenotype - Transverse Colonic Stricture" equaling "Yes" are
 considered evidence of disease activity in the colon.}
 \item{ Any record of Phenotype - Esophageal Stricture",
 "Phenotype - Jejunal Stricture", "Phenotype - Duodenal Stricture" or
 "Phenotype - Gastric Stricture " equal to "Yes" will be considered evidence
 of disease activity in the upper GI tract.}
 }
}
