% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_manitoba}
\alias{calculate_manitoba}
\title{calculate_manitoba}
\usage{
calculate_manitoba(observations)
}
\arguments{
\item{observations}{observation table usually uploaded using load_data}
}
\value{
A dataframe with all manitoba from eCRF  regardless of IBD diagnosis.
}
\description{
Calculates manitoba score from SPARC data.
}
