% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_mes}
\alias{calculate_mes}
\title{calculate_mes}
\usage{
calculate_mes(procedures)
}
\arguments{
\item{procedures}{procedures table usually uploaded using load_data}
}
\value{
A dataframe with all mes scores that could be calculated regardless of IBD diagnosis.
}
\description{
Calculates Mayo Endoscopy score from SPARC data
}
