% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_pga}
\alias{calculate_pga}
\title{calculate_pga}
\usage{
calculate_pga(observations)
}
\arguments{
\item{observations}{observation table usually uploaded using load_data}
}
\value{
A dataframe with all PGA from Smartform  regardless of IBD diagnosis.
}
\description{
Calculates PGA score from SPARC data.
}
