% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_pro2}
\alias{calculate_pro2}
\title{calculate_pro2}
\usage{
calculate_pro2(observations)
}
\arguments{
\item{observations}{observation table usually uploaded using load_data}
}
\value{
A dataframe with all pro2 scores from eCRF and Smartform regardless of IBD diagnosis.
}
\description{
Calculates pro2 from SPARC data.
}
\details{
remission/mild/moderate/severe definition from https://pubmed.ncbi.nlm.nih.gov/25348809/
}
