% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_pro3}
\alias{calculate_pro3}
\title{calculate_pro3}
\usage{
calculate_pro3(observations)
}
\arguments{
\item{observations}{observation table usually uploaded using load_data}
}
\value{
A dataframe with all pro3 scores from eCRF and Smartform regardless of IBD diagnosis.
}
\description{
Calculates pro3 from SPARC data.
}
