% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_ses}
\alias{calculate_ses}
\title{calculate_ses}
\usage{
calculate_ses(procedures)
}
\arguments{
\item{procedures}{procedures table usually uploaded using load_data}
}
\value{
A dataframe with all ses scores that could be calculated regardless of IBD diagnosis.
}
\description{
Calculates SES-CD score from SPARC data
}
