% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_calculations.R
\name{calculate_uc_phenotype}
\alias{calculate_uc_phenotype}
\title{calculate_uc_phenotype}
\usage{
calculate_uc_phenotype(observations, worsening = FALSE)
}
\arguments{
\item{observations}{observations table usually generated using load_data}

\item{worsening}{a logical indicating whether the worst extend of ulcerative colitis up to a specific date should be returned.}
}
\value{
A dataframe with all ulcerative colitis phenotypes that could be
derived from the data regardless of IBD diagnosis.
}
\description{
Calculates ulcerative colitis phenotypes from SPARC data
}
\details{
If two or more records of ulcerative colitis phenotypes are available
for a patient on a single day the record showing the larger extend of
ulcerative colitis is chosen, with Pancolitis > Extensive ulcerative colitis >
Left-sided ulcerative colitis > Ulcerative proctitis > Unknown.
}
