% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculations.R
\name{calculate_ucdai}
\alias{calculate_ucdai}
\title{calculate_ucdai}
\usage{
calculate_ucdai(observations)
}
\arguments{
\item{observations}{observation table usually uploaded using load_data}
}
\value{
A dataframe with all 6pt & 9pt scores from eCRF and Smartform regardless of IBD diagnosis.
}
\description{
Calculates 6pt & 9pt UCDAI scores from SPARC data.
}
