% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_categorization.R
\name{categorize_race}
\alias{categorize_race}
\title{categorize race}
\usage{
categorize_race(race)
}
\arguments{
\item{race}{the data frame generated by extract_race function}
}
\value{
A dataframe with categorized race & ethnicity information stored in column RACE_CAT and ETHNICITY_CAT.
}
\description{
Categorize race & ethnicity into minimum choices from Collection of Race and Ethnicity Date in Clinical Trials and Clinical Studies for FDA-Regulated Medication Products
https://pink.citeline.com/-/media/supporting-documents/pink-sheet/2024/01/race_ethnicity_data_guidance_01292024.pdf?rev=1b3d813f6d254080bae6ec3e6ab872d0&hash=C485BD27F152B7DEBB819B012A644178&utm_campaign=The%20Evidence%20Digest&utm_medium=email&_hsmi=296353360&_hsenc=p2ANqtz-9lXtx6rO5GP2zHvxvxL4TwfKjSw5KZtP7MxmnXiPWJjaYt6F1LrPLIkiWqX1ynqmmFwtcgwsqewXuX5I25Dz7B5pi2Z6nYr_mmXnSEO0xIBNfWL6A&utm_content=296353360&utm_source=hs_email
EMR and eCRF data sources are used
}
