% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cd_codes}
\alias{cd_codes}
\title{cd_codes}
\format{
A data frame with 40 rows and 4 variables:
\describe{
  \item{Phenotype}{Crohn's Disease Phenotype}
  \item{CODING}{Yes, No or Unknown}
  \item{OBS_TEST_CONCEPT_CODE}{Corresponding Concept Code for each coding}
  \item{TIER}{Level of importance for data capture}

}
}
\usage{
cd_codes
}
\description{
A dataframe containing the coding for Crohn's Disease Phenotypes
}
\keyword{datasets}
