% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_med_journey_helpers.R
\name{clb_current_med_dates}
\alias{clb_current_med_dates}
\title{clb_current_med_dates}
\usage{
clb_current_med_dates(medication, encounter)
}
\arguments{
\item{medication}{A dataframe with all medication data created using sparc_med_filter.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}
}
\value{
A dataframe with the first medication start date for each drug.
}
\description{
current_med_dates function for CLB version of SPARC med journey
}
