% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_patients.R
\name{count_patients}
\alias{count_patients}
\title{count_patients}
\usage{
count_patients(data)
}
\arguments{
\item{data}{A dataframe containing SPARC/Qorus data of any type.}
}
\value{
Numeric result containing number of patients
}
\description{
Count unique Sparc and Qorus Patients in a dataframe using unique MPI.
}
