% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning_functions.R
\name{crp_active_clean}
\alias{crp_active_clean}
\title{crp_active_clean}
\usage{
crp_active_clean(crp, labs)
}
\arguments{
\item{crp}{The table of CRP labs created by extract_lab_emr.}

\item{labs}{The table of labs from the ddm extract}
}
\value{
Table with active or inactive flag for CRP level with an index date.
}
\description{
Function that cleans and creates a flag for active level for CRP.
}
