% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{current_med}
\alias{current_med}
\title{current_med}
\usage{
current_med(medication)
}
\arguments{
\item{medication}{A dataframe with generated by sparc_med_filter}
}
\value{
A dataframe with all current medications in eCRF (excludes not current)
}
\description{
Find current medication in eCRF data source
}
