% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_escalation.R
\name{dose_escalation}
\alias{dose_escalation}
\title{dose_escalation}
\usage{
dose_escalation(medication)
}
\arguments{
\item{medication}{A dataframe generated from sparc_med_filter()}
}
\value{
A dataframe with the master patient id, data source, medication name, original dose and any doses reported over time for that source.
}
\description{
For SPARC, this function finds if a medication has increased in dose over time.
}
