% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eim_codes}
\alias{eim_codes}
\title{eim_codes}
\format{
A data frame with 21 rows and 3 variables:
\describe{
  \item{EIM_DX}{Extra Intestional Manifestation}
  \item{CODING}{Yes/No or Unknown}
  \item{DIAG_CONCEPT_CODE}{Corresponding Concept Code for each coding}

}
}
\usage{
eim_codes
}
\description{
A dataframe containing the coding for Extra Intestinal Manifestations
}
\keyword{datasets}
