% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_emr_extract.R
\name{emr_extract_diagnosis}
\alias{emr_extract_diagnosis}
\title{emr_extract_diagnosis}
\usage{
emr_extract_diagnosis(
  data,
  text_search = F,
  code_search = T,
  inclusion = NULL,
  exclusion = NULL,
  custominc = NULL,
  customexc = NULL,
  datecutoff = NULL,
  index_range = 36500
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, and the diagnosis AND/OR patient problem tables. Only EMR data needs to be uploaded.}

\item{text_search}{Option to extract relevant diagnosis codes using keyword text search. Defaults to false.}

\item{code_search}{Option to extract relevant diagnosis codes using only ICD 10 codes. Defaults to true.}

\item{inclusion}{When searching by code, the diagnosis category of ICD 10 codes to search for. Can choose between predefined categories (included codes are detailed in Diagnosis_EMR_Extract.Rd) or enter a custom defined list. When text searching, a regular expression string of keywords to search for (eg. "argument1|argument2|argument3")}

\item{exclusion}{diagnosis category of ICD 10 codes to exclude from search. Will be input as custom list, may add predefined lists in future updates.}

\item{custominc}{optional list of custom ICD 10 codes to be searched for. The variable "inclusion" must also be set to "CUSTOM" for this option. If multiple inputs, needs to be in a regular expression string (eg. "argument1|argument2|argument3")}

\item{customexc}{optional list of custom ICD 10 codes to be excluded from the search. The variable "exclusion" must also be set to "CUSTOM" for this option. If multiple inputs, needs to be in a regular expression string (eg. "argument1|argument2|argument3")}

\item{datecutoff}{Type of date cutoff to be used for EMR extract. Current options include "ENROLLMENT" - number of days before consent into study.}

\item{index_range}{Number of days before date cutoff to be used for EMR extract. Time must be specified in number of days. Default is 36500 days (100 years).}
}
\value{
A list of dataframes with all relevant EMR data.
}
\description{
A function to extract relevant diagnosis codes from
diagnosis and/or patient problem EMR data for a specific condition of
interest.
}
\examples{

# example with predefined category (all cancer diagnoses excluding skin carcinomas)

# cancerdx_woutskin = diagnosis_emr_extract(data, inclusion="CANCER", exclusion = "SKIN CARCINOMA")

# example with custom category (all cancer diagnoses excluding cervical carcinomas)
#example with predefined category

#cancerdx_woutskin = diagnosis_emr_extract(data, inclusion="CANCER", exclusion = "SKIN CARCINOMA")

#example with custom category

#cancerdx_woutcervical=emr_extract_diagnosis(data,inclusion="CUSTOM", exclusion = "CUSTOM",custominc="c|d0|d1|d2|d3|d4",customexc="d06")

#example of text search only
#textonly_abscess <- emr_extract_diagnosis(data, text_search = T, code_search = F, inclusion = "abscess")
}
