% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr_loading_dose.R
\name{emr_loading_pattern}
\alias{emr_loading_pattern}
\title{emr_loading_pattern}
\usage{
emr_loading_pattern(prescriptions, encounter)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}

\item{encounter}{a dataframe with encounter data usually generated using load_data.}
}
\value{
A dataframe with the first medication start date for each drug. Bionaive is 1 if a patient has no prior reported biologics (including JAK inhibitors and Tofacitinib). Started after enrollment is 1 if the medication start date is after the date of consent.
}
\description{
Find if emr prescriptions follows known loading patterns for biologics
}
