% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{extract_consent}
\alias{extract_consent}
\title{extract_consent}
\usage{
extract_consent(demographics, study)
}
\arguments{
\item{demographics}{A dataframe with demographics data.}

\item{study}{Cohort from which the consent data should be extracted. Either SPARC or QORUS.}
}
\value{
A dataframe with consent data.
}
\description{
Extract consent data for SPARC & QORUS from a table with demographic information.
}
\details{
When a patient withdraws and re-enrolls, the first date of consent is used.
}
