% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{extract_demo}
\alias{extract_demo}
\title{extract_demo}
\usage{
extract_demo(demographics, study)
}
\arguments{
\item{demographics}{A dataframe with demographics data.}

\item{study}{Cohort from which the data should be extracted. Either SPARC or QORUS.}
}
\value{
A dataframe with birth year and sex information.
}
\description{
Extract Birth Year and Sex for SPARC & QORUS from a table with demographic information.
}
\details{
Preference is given to the EMR data source when available
}
