% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{extract_diagnosis}
\alias{extract_diagnosis}
\title{extract_diagnosis}
\usage{
extract_diagnosis(diagnosis, encounter, demographics, study)
}
\arguments{
\item{diagnosis}{A dataframe with diagnosis data.}

\item{encounter}{A dataframe with encounter data. Only required for SPARC.}

\item{demographics}{A dataframe with demographics data.}

\item{study}{Cohort for which diagnosis information is needed. Either SPARC or QORUS.}
}
\value{
A dataframe with diagnosis data.
}
\description{
Extract IBD diagnosis and diagnosis date for SPARC & QORUS from a table with diagnosis data.
}
