% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_of_interest.R
\name{extract_endoscopy}
\alias{extract_endoscopy}
\title{extract_endoscopy}
\usage{
extract_endoscopy(procedures)
}
\arguments{
\item{procedures}{A dataframe with procedures data.}
}
\value{
A dataframe with endoscopy date and indication data.
}
\description{
Extract endoscopy data for SPARC.
}
