% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_emr_extract.R
\name{extract_lab_emr}
\alias{extract_lab_emr}
\title{lab_emr_extract}
\usage{
extract_lab_emr(
  data,
  labtype,
  codetype = c("LOINC", "KEYWORDS", "ALL"),
  positive_elevated = FALSE,
  datecutoff = NULL,
  index_range = 36500
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, and the lab tables. Only EMR data needs to be uploaded.}

\item{labtype}{The type of lab results to select ("CRP","HS CRP","NS CRP","FCAL"). If the user inputs any other value, the lab EMR table will be returned.}

\item{codetype}{The criteria to use for querying codes. The options are "LOINC" for only related LOINC codes,"KEYWORDS" for only related words in the Code Description field, or "ALL" for both critera.}

\item{positive_elevated}{If "TRUE", the results will only be for results above a "positive" and/or "elevated" threshold, dependent on the test category.}

\item{datecutoff}{Type of date cutoff to be used for EMR extract. Current options include "BEFORE ENROLLMENT" or "AFTER ENROLLMENT" - number of days before or after consent into study.}

\item{index_range}{Number of days before date cutoff to be used for EMR extract. Time must be specified in number of days. Default is 36500 days (100 years).}
}
\value{
A list of dataframes with all relevant EMR lab result records
}
\description{
A function to extract relevant lab results from
lab EMR data for a specific test of interest.
}
\details{
1. user defines type of lab results to select (CRP, HS CRP, NS CRP, FCAL). Alb blood/serum, Glucose blood, Glucose urine to be added in a future update.
2. user defines blood, serum, urine, stool, etc (only for specific types of tests and will be added in a future update.)
3. user defines if searching for only LOINC codes, all codes or key words as well
4. user defines if all or positive results (default "positive/elevated" definition or custom(add later if possible))
5. user defines if only valid units for positive results or all positive results
}
\examples{


#example for extracting all C-Reactive Protein lab results
#labdomemr = load_data(datadir="C:/Users/me/",domains=c("Labs","Demographics"),data_type="EMR")
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL")

#example for extracting all High Sensitivity C-Reactive Protein lab results
#hscrp = extract_lab_emr(data=labdomemr,labtype="HS CRP",codetype="ALL")

#example for extracting all non-High Sensitivity C-Reactive Protein lab results
#nscrp = extract_lab_emr(data=labdomemr,labtype="NS CRP",codetype="ALL")

#example for extracting all Fecal Calprotectin lab results
#fcal = extract_lab_emr(data=labdomemr,labtype="FCAL",codetype="ALL")

#example for extracting LOINC C-Reactive Protein lab results
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="LOINC")

#example for extracting C-Reactive Protein lab results using only keywords
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="KEYWORDS")

#example for extracting all elevated C-Reactive Protein lab results
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",positive_elevated=TRUE)
#example for extracting all C-Reactive Protein lab results from after patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT")

#example for extracting all C-Reactive Protein lab results from one year after patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT",index_range=365)

#example for extracting all C-Reactive Protein lab results from two years after patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT",index_range=730)

#example for extracting all C-Reactive Protein lab results from before patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT")

#example for extracting all C-Reactive Protein lab results from one year before patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT",index_range=365)

#example for extracting all C-Reactive Protein lab results from two years before patients were enrolled in the respective study
#crp = extract_lab_emr(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT",index_range=730)



}
