% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{extract_labs}
\alias{extract_labs}
\title{extract_labs}
\usage{
extract_labs(labs, test)
}
\arguments{
\item{labs}{A dataframe with the labartory results in it. This can be loaded using load_data(datadir = "data/", cohort = "SPARC", domains = c("labs"), data_type = "CRF")}

\item{test}{The test of interest - either fcal for fecal calprotectin or hscrp for High-sensitivity C-reactive Protein.}
}
\value{
A dataframe with the master patient id, lab result, units and date of specimen collection
}
\description{
For the hsCRP analysis done by the central lab, the Beckman Coulter analyzer and calibration kit was used (https://mms.mckesson.com/product/727502/Beckman-Coulter-ODR3021). For the
fecal calprotectin analysis we used the Buhlmann fCAL ELISA (https://buhlmannlabs.com/buhlmann-fcal-elisa/#laboratory).
}
\details{
Filter lab crf data for the hsCRP or fecal calprotectin lab results performed by the Crohn's & Colitis Foundation on biosamples collected for SPARC.
}
