% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformations.R
\name{extract_race}
\alias{extract_race}
\title{extract_race}
\usage{
extract_race(demographics, study)
}
\arguments{
\item{demographics}{A dataframe with demographics data.}

\item{study}{Cohort from which the data should be extracted. Either SPARC or QORUS.}
}
\value{
A dataframe with birth year and sex information.
}
\description{
Extract race & ethnicity information SPARC & QORUS from a table with demographic information.
EMR and eCRF data sources are used
}
