% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyibdhist_emr_patient.R
\name{familyibdhist_emr_patient}
\alias{familyibdhist_emr_patient}
\title{familyibdhist_emr_extract}
\usage{
familyibdhist_emr_patient(
  data,
  relation = NULL,
  negativemention = FALSE,
  famhistnoibd = FALSE,
  nofamhist = FALSE
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, and the Patient History tables. Only EMR data needs to be uploaded.}

\item{relation}{If "FIRST_DEG", create an indicator variable (ibdfamhistfirst) specifying which patients have IBD family history records with first degree family members.}

\item{negativemention}{If "TRUE", create an indicator variable (noibdfamhist) specifying which patients have records specifically citing NO family history of IBD. The default is FALSE.}

\item{famhistnoibd}{If "TRUE", create an indicator variable (famhistnoibd) specifying which patients have family history EMR but no mention of IBD. The default is FALSE.}

\item{nofamhist}{If "TRUE".create an indictor variable (nofamhist) specifying which patients do not have any records in the family history EMR table}
}
\value{
A modified patient list with all specified indicators.
}
\description{
A function to add an indicator variable to a patient table for patients with records from the Family History EMR table indicating a family member with a history of IBD.
User can add up to five indicator variables:
1. Patients with any family members with IBD (ibdfamhist)
2. Only First Degree Family members with IBD. First degree is defined as: Biological Parent, Sibling or Child. (ibdfamhistfirst)
3. Patients with records indicating "NO" family members with IBD. (These patients are excluded in the first two options) (noibdfamhist)
4. Patients with family history EMR information but no indication of IBD history. (famhistnoibd)
5. Patients with no family history EMR information. (nofamhist)
}
\examples{

#example for generating a patient level data frame with an indicator variable for patient family history of IBD

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,relation=NULL,negativemention=FALSE,famhistnoibd=FALSE,nofamhist=FALSE)
#patient = familyibdhist_emr_patient(data)
#example for generating a patient level data frame with indicator variables for patient family history of IBD and for only first degree family member history of IBD

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,relation="FIRST_DEG")
#example for generating a patient level data frame with indicator variables for patient family history of IBD and for records of no IBD family history

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,negativemention=TRUE)
#example for generating a patient level data frame with indicator variables for patient family history of IBD and for family history EMR but no mention of IBD

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,famhistnoibd=TRUE)
#example for generating a patient level data frame with indicator variables for patient family history of IBD and for no family history EMR

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,nofamhist=TRUE)
#example for generating a patient level data frame with all five indicator variables.

#data = load_data(datadir="C:/Users/me/",domains=c("Patient_History","Demographics"),data_type="EMR")
#patient = familyibdhist_emr_patient(data,relation="FIRST_DEG",negativemention=TRUE,famhistnoibd=TRUE,nofamhist=TRUE)
}
