% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cleaning_functions.R
\name{fcal_active_clean}
\alias{fcal_active_clean}
\title{fcal_active_clean}
\usage{
fcal_active_clean(fcal, labs)
}
\arguments{
\item{fcal}{The table of FCAL labs created by extract_lab_emr}

\item{labs}{The ddm extract with the labs table}
}
\value{
Table with active or inactive flag for FCAL level with an index date.
}
\description{
Function that cleans and creates a flag for active level for fecalcalprotectin.
}
