% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fix_col_names}
\alias{fix_col_names}
\title{fix_col_names}
\usage{
fix_col_names(df)
}
\arguments{
\item{df}{The dataframe created for the summary table}
}
\value{
The dataframe with corrected names
}
\description{
standardize column names for summary tables
}
