% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{folder_fix}
\alias{folder_fix}
\title{folder_fix}
\usage{
folder_fix(folder)
}
\arguments{
\item{folder}{The folder specified for the function}
}
\value{
the folder name in the correct format
}
\description{
add "/" to end of folder name for functions
}
