% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_change.R
\name{frequency_change}
\alias{frequency_change}
\title{frequency_change}
\usage{
frequency_change(medication)
}
\arguments{
\item{medication}{A dataframe generated from sparc_med_filter()}
}
\value{
A dataframe with the master patient id, data source, medication name, original prescribed interval and changing frequency
}
\description{
For SPARC, this function finds if a long term advanced therapy has changed in frequency over time.
}
