% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdai.R
\name{full_pcdai}
\alias{full_pcdai}
\title{full_pcdai}
\usage{
full_pcdai(observations, encounter, labs)
}
\arguments{
\item{observations}{A dataframe with observation data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}

\item{labs}{A dataframe with lab data usually generated using load_data.}
}
\value{
A dataframe with PCDAI scores
}
\description{
A function that creates a table with the available full PCDAI scores for RISK patients.
}
