% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_emr_patient.R
\name{lab_emr_patient}
\alias{lab_emr_patient}
\title{lab_emr_patient}
\usage{
lab_emr_patient(
  data,
  labtype,
  codetype = c("LOINC", "KEYWORDS", "ALL"),
  positive_elevated = FALSE,
  datecutoff = NULL,
  index_range = 36500
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, and the lab tables. Only EMR data needs to be uploaded.}

\item{labtype}{The type of lab results to select ("CRP","HS CRP","NS CRP","FCAL"). If the user inputs any other value, the lab EMR table will be returned.}

\item{codetype}{The criteria to use for querying codes. The options are "LOINC" for only related LOINC codes,"KEYWORDS" for only related words in the Code Description field, or "ALL" for both critera.}

\item{positive_elevated}{If "TRUE", the results will only be for results above a "positive" and/or "elevated" threshold, dependent on the test category.}

\item{datecutoff}{Type of date cutoff to be used for EMR extract. Current options include "BEFORE ENROLLMENT" or "AFTER ENROLLMENT" - number of days before or after consent into study.}

\item{index_range}{Number of days before date cutoff to be used for EMR extract. Time must be specified in number of days. Default is 36500 days (100 years).}
}
\value{
A patient list of with all relevant EMR lab result indicators
}
\description{
A function to add an indicator variable to a patient table for patients with records from the Lab EMR table indicating the presence of the specified lab test ("CRP","HS CRP","NS CRP","FCAL").
1. user defines type of lab results to select (CRP, HS CRP, NS CRP, FCAL). Alb blood/serum, Glucose blood, Glucose urine to be added in a future update.
2. user defines blood, serum, urine, stool, etc (only for specific types of tests and will be added in a future update.)
3. user defines if searching for only LOINC codes, all codes or key words as well
4. user defines if all or positive results (default "positive/elevated" definition or custom(add later if possible))
5. user defines if only valid units for positive results or all positive results
}
\details{
User can add up to two indicator variables:
1. Patients with the specified lab test. This indicator is included with all results. The user can specify if the results need to be within a specified time frame. (labdataall)
2. Patients with elevated or positive results for the specified lab test. This indicator is optional. The user can specify if the results need to be within a specified time frame. (labdatapos)

Future Enhancements:
3. Patients with lab EMR information but no record (within the time frame if specified) for the test. (This will be added on a future date) (labdatanotest)
4. Patients with no lab EMR information. (This will be added on a future date) (nofamhist)
}
\examples{



#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results
labdomemr = load_data(datadir="C:/Users/me/",domains=c("Labs","Demographics"),data_type="EMR")
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL")

#example for creating a patient level dataset with an indicator for all High Sensitivity C-Reactive Protein lab results
patient = lab_emr_patient(data=labdomemr,labtype="HS CRP",codetype="ALL")

#example for creating a patient level dataset with an indicator for all non-High Sensitivity C-Reactive Protein lab results
patient = lab_emr_patient(data=labdomemr,labtype="NS CRP",codetype="ALL")

#example for creating a patient level dataset with an indicator for all Fecal Calprotectin lab results
patient = lab_emr_patient(data=labdomemr,labtype="FCAL",codetype="ALL")

#example for creating a patient level dataset with an indicator for LOINC C-Reactive Protein lab results
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="LOINC")

#example for creating a patient level dataset with an indicator for C-Reactive Protein lab results using only keywords
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="KEYWORDS")

#example for creating a patient level dataset with an indicator for all elevated C-Reactive Protein lab results and all C-Reactive Protein lab results
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",positive_elevated=TRUE)
#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from after patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT")

#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from one year after patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT",index_range=365)

#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from two years after patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="AFTER ENROLLMENT",index_range=730)

#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from before patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT")

#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from one year before patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT",index_range=365)

#example for creating a patient level dataset with an indicator for all C-Reactive Protein lab results from two years before patients were enrolled in the respective study
patient = lab_emr_patient(data=labdomemr,labtype="CRP",codetype="ALL",datecutoff="BEFORE ENROLLMENT",index_range=730)


}
