% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_result_cln.R
\name{lab_result_cln}
\alias{lab_result_cln}
\title{lab_result_cln}
\usage{
lab_result_cln(labemrdata)
}
\arguments{
\item{data}{A single dataframe with lab EMR records, from a list of dataframes, typically generated by a call to load_data. See example for extracting the lab EMR table from the load_data results.}
}
\value{
A modified patient list with all specified indicators.
}
\description{
A function to clean the "Test Unit" and "Lab Results" fields in in the Lab EMR table to standard formats for analysis.
}
\examples{

#example for cleaning a lab EMR table starting with a list of dataframes, typically generated by a call to load_data

data = load_data(datadir="C:/Users/me/",domains="Labs",data_type="EMR")
#labemr = data$labs
#cleanedlabemr = lab_result_cln(labdata)
}
