% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plexus_data_loading_redshift.R
\name{load_data_redshift}
\alias{load_data_redshift}
\title{load_data_Redshift_AZ}
\usage{
load_data_redshift(
  conn,
  cohort = c("RISK", "QORUS", "SPARC"),
  domains = c("ALL"),
  data_type = c("BOTH", "CRF", "EMR")
)
}
\arguments{
\item{conn}{A connection to a redshift instance. Tables are assumed to be stored under a schema called "ibd_plexus"}

\item{cohort}{The cohort to load. Either RISK, QORUS, or SPARC.}

\item{domains}{The domains to load. Default is "All". Must be a character string.}

\item{data_type}{The data source to load either case report forms, electronic medical record or both. Options are both, crf or emr.}
}
\value{
A list of dataframes for each domain. If both sources are loaded, emr and crf data are combined.
}
\description{
Loads data from an AWS Reshift instance.
}
\details{
When the index date is the date of consent, only the patient reported data is used.
}
