% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plexus_data_loading_script.R
\name{load_zipped_data}
\alias{load_zipped_data}
\title{load_zipped_data}
\usage{
load_zipped_data(
  datadir,
  cohort = c("RISK", "QORUS", "SPARC", "SIRQC"),
  domains = c("ALL"),
  data_type = c("BOTH", "CRF", "EMR"),
  exdir = "."
)
}
\arguments{
\item{datadir}{The directory where data is saved.}

\item{cohort}{The cohort to load. Either RISK, QORUS, SPARC or SIQRC.}

\item{domains}{The domains to load. Default is "All". Must be a character string.}

\item{data_type}{The data source to load either case report forms, electronic medical record or both. Options are both, crf or emr.}

\item{exdir}{The file to save the files once unzipped. If left blank will be the working directory. Do not include backslash at the end of the file location.}
}
\value{
A list of dataframes for each domain. If both sources are loaded, emr and crf data are combined.
}
\description{
Load compressed files extracted from IBD Plexus.
}
