% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_grp}
\alias{med_grp}
\title{med_grp}
\format{
A data frame with 175 rows and 3 variables:
\describe{
  \item{MEDICATION_NAME}{Medication Name as it appears in SF, ECRF or EMR}
  \item{med_type}{The class of medication. Medication classes include: Aminosalicylates, antibiotics, antidiarrheals, biologics, corticosteroids, immunomodulators, probiotics, and targeted synthetic small molecules}
  \item{new_med_name}{Standardized medication name}
  \item{biosimilar_flag}{Equals 1 for biosimilars}

}
}
\usage{
med_grp
}
\description{
A dataframe containing the coding for Medications. Allows to group the same
medications together if they have different names (generic v. scientific)
from different source.
}
\details{
Infliximab Biosimilars added March 2022.
Adalimumab Biosimilar added December 2022.
Additional Adalimumab Biosimilars added July 2023.
Steroid list expanded July 2023.
New targeted synthetic small molecules med_type added.
Ustekinumab and other Biosimilars added August 2025.
}
\keyword{datasets}
