% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{med_search}
\alias{med_search}
\title{med_search}
\usage{
med_search(df, med_name, med_string)
}
\arguments{
\item{df}{scripts data frame to filter through}

\item{med_name}{list of medication names to search for}

\item{med_string}{column of medication to search through}
}
\value{
filtered data frame with medications of interest
}
\description{
filters prescription tables for medications of interest
}
