% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{med_verification}
\alias{med_verification}
\title{med_verification}
\usage{
med_verification(medication)
}
\arguments{
\item{medication}{medication data frame}
}
\value{
data frame with the medication verification date
}
\description{
Adds most recent medication verification date to medication journey.
}
