% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_med_journey_helpers.R
\name{overlap}
\alias{overlap}
\title{overlaps}
\usage{
overlap(med)
}
\arguments{
\item{med}{A dataframe med created halfway through the sparc med journey function}
}
\value{
A dataframe with the medication overlaps for each drug.
}
\description{
current_med_dates function for CLB version of SPARC med journey
}
