% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_med_journey_helper.R
\name{overlapping_meds}
\alias{overlapping_meds}
\title{overlapping_meds}
\usage{
overlapping_meds(table)
}
\arguments{
\item{table}{A dataframe created by the first part of the risk medication journey function.}
}
\value{
A dataframe of RISK patients with additional columns of the overlapping medications and the days the medications overlapped
}
\description{
create table with overlapping meds column and overlapping med days columns for RISK patients
}
