% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qorus_summary_at_index.R
\name{qorus_summary}
\alias{qorus_summary}
\title{Create QORUS Summary Table - all responses}
\usage{
qorus_summary(
  data,
  index_info = c("ENROLLMENT", "LATEST", "ENDOSCOPY", "OMICS"),
  filename = "QORUS_SUMMARY.xlsx",
  index_range = "60",
  location_logic = "CCF",
  export = T
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, and observations. Only CRF data needs to be uploaded.}

\item{index_info}{A dataframe  with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Other options include, enrollment, latest, endoscopy, omics collection.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{index_range}{the number of days to look out from index date. If multiple of the same score fall within this range then the one closest to the index date is picked. Default is 60 days.}

\item{location_logic}{specify which CD location function to use. If using logic developed by AZ then use location_logic = "AZ".}

\item{export}{if TRUE then writes the data frame to its own excel file}
}
\value{
qorus summary table
}
\description{
Create QORUS Summary Table - all responses
}
