% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
data from csv/txt files
}
\description{
Function to read data tables into a list and combine tables of each type. From rancho.
}
