% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{reason_stopped}
\alias{reason_stopped}
\title{reason_stopped}
\usage{
reason_stopped(prescriptions)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}
}
\value{
A dataframe with the reason stopped
}
\description{
Extract reason stopped for ECRF and SF SPARC Data source
}
