% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{remove_empty_cols}
\usage{
remove_empty_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
non empty columns
}
\description{
Function to remove empty columns from a table. From rancho.
}
