% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_emr_extract.R
\name{return_tables}
\alias{return_tables}
\title{return_tables}
\usage{
return_tables(data, dxemricd = NULL, ppemricd = NULL, phemricd = NULL)
}
\arguments{
\item{data}{The data frame that is inputted into the emr_extract_diagnosis function}

\item{dxemricd}{filtered diagnosis data frame from emr_extract_diagnosis function}

\item{ppemricd}{filtered patient_problem data frame from emr_extract_diagnosis function}

\item{phemricd}{filtered patient_history data frame from emr_extract_diagnosis function}
}
\value{
a list of the data frames available that are filtered using the emr_extract_diagnosis function
}
\description{
A function that returns the filtered dataframes that were imputed into extract_emr_diagnosis function. Can be any combination of diagnosis, patient_problem and/or patient_history.
}
