% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_med_journey_helper.R
\name{risk_antibiotic_rounds}
\alias{risk_antibiotic_rounds}
\title{risk_antibiotics_rounds}
\usage{
risk_antibiotic_rounds(prescriptions, encounter)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions for RISK data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}
}
\value{
A dataframe of RISK patients the rounds of antibiotics they were on
}
\description{
create table with antibiotics rounds for RISK patients
}
