% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clinical_data_v2.R
\name{risk_extract_clinical_data}
\alias{risk_extract_clinical_data}
\title{risk_extract_clinical_data}
\usage{
risk_extract_clinical_data(
  dir,
  select.col = c("DIAG_CONCEPT_NAME", "DIAGNOSIS_DATE", "GENDER", "TYPE_OF_ENCOUNTER",
    "VISIT_ENCOUNTER_START_DATE", "AGE_AT_ENCOUNTER", "BIRTH_YEAR",
    "LAB_TEST_CONCEPT_NAME", "ASSAY.NAME", "MEDICATION_NAME", "OBS_TEST_CONCEPT_NAME",
    "ANA_SITE_CONCEPT_NAME", "BIOSAMPLE_CONCEPT_NAME"),
  select.val = c("DIAG_STATUS_CONCEPT_NAME", "DIAGNOSIS_DATE", "GENDER",
    "TYPE_OF_ENCOUNTER", "VISIT_ENCOUNTER_START_DATE", "AGE_AT_ENCOUNTER", "BIRTH_YEAR",
    "TEST_RESULT_NUMERIC", "RAW.DATA.FILE.NAME", "MED_START_DATE", "TEST_RESULT_NUMERIC",
    "TEST_RESULT_NUMERIC", "Sample.Status"),
  filter.col = c("DIAG_CONCEPT_NAME", "OBS_TEST_CONCEPT_NAME"),
  filter.val = c("IBD - Family History", "Disease Location",
    "Endoscopic Assessment - Deep Ulceration",
    "Endoscopic Assessment - Superficial Ulceration",
    "Endoscopic Assessment - Amount of Surface Ulcerated",
    "Endoscopic Assessment - Amount of Surface Involved", "Perianal Disease -", "EIM",
    "Disease Behavior - Stricturing/Fibrostenotic",
    "Disease Behavior - Internally Pentrating", "PCDAI"),
  filename = "RISK_Summary.xlsx"
)
}
\arguments{
\item{dir}{The directory where the unzipped data is saved.}

\item{select.col}{Columns to select.}

\item{select.val}{Columns with values for the selected columns. If select.val
is the same as the select.col, then the whole column will be selected. If
value is different than the select.col the selected column will be reshaped
into a wide format with values from the select.val column.}

\item{filter.col}{Columns that need to be filtered prior to reshaping.}

\item{filter.val}{Values to filter on filter.col.}

\item{filename}{A filename to write the final output to.}
}
\value{
dataframe and wide excel file.
}
\description{
version 1.3: February 11, 2019
IBD Plexus, Crohn's & Colitis Foundation
Created by Victoria Zadorozhny, Rancho Biosciences
}
\details{
The following script reads RISK study clinical data tables from the
specified input directory, extracts specified columns and values,
calculates wPCDAI score, orders columns and outputs
a summary table as an excel workbook.
}
