% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_med_journey.R
\name{risk_med_journey}
\alias{risk_med_journey}
\title{risk_med_journey}
\usage{
risk_med_journey(prescriptions, encounter)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions for RISK data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}
}
\value{
An table with the RISK medication journey table.

#' @details Medication start and stop takes are from RISK case report forms. If there is no medication end date, but a new medication with the same mechanism of action was started, the day before the start of the new medication is used as the medication end date.

The following columns are convenience flags and indices to facilitate easy filtering of the data:

\describe{
  \item{MEDICATION_NUMBER}{counts the number of different medications in a patients' journey.
The medication with with earliest start date will have MEDICATION_NUMBER = 1.}
  \item{STEROID_OVERLAP}{is 1 if a patient also reported being on steroids at some point that overlaps with the medication interval.}
  \item{ANTIBIOTIC_OVERLAP}{is 1 if a patient also reported being on an antibiotic at some point that overlaps with the medication interval.}
}
}
\description{
Creates RISK medication journey table for long term medications including Biologics, 5-ASA's and Immunomodulators.
}
\examples{
# upload data
# dat <- load_data("~/r_input/", cohort = "RISK", domains = c("prescriptions", "encounter"), data_type = "CRF")
# export RISK Summary Table to xlsx
# risk_med_journey_df <- risk_med_journey(dat$prescriptions, dat$encounter)



}
