% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{risk_meds_at_visit}
\alias{risk_meds_at_visit}
\title{risk_meds_at_visit}
\usage{
risk_meds_at_visit(prescriptions, encounter)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}
}
\value{
A dataframe of RISK patients with a column MEDICATIONS_AT_VISIT, deidentified patient and encounter ID's, and date columns
}
\description{
create table with column for RISK patients that lists the medications they were on at a visit
}
