% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_summary.R
\name{risk_summary}
\alias{risk_summary}
\title{risk_summary}
\usage{
risk_summary(dir, data, filename = "RISK_Summary.xlsx")
}
\arguments{
\item{dir}{The directory to where data from IBD Plexus is saved.}

\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, prescriptions and labs.}

\item{filename}{the name of the output file. Must be .xlsx.}
}
\value{
An excel file with the RISK Summary Table.
}
\description{
Creates RISK Cohort Summary Table with wPCDAI calculations, and exports as an excel sheet
}
\examples{
# upload data
# dat <- load_data(dir, cohort = "RISK", domains = c("ALL"), data_type = "BOTH")
# export RISK Summary Table to xlsx
# risk_summary(dir, data = dat, filename = "RISK_Summary.xlsx")


}
