% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_biosample_flags.R
\name{sparc_biosample_flags}
\alias{sparc_biosample_flags}
\title{sparc_biosample_flags}
\usage{
sparc_biosample_flags(data, index_range)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data.
Must include demographics, diagnosis, encounter, procedures, observations,
biosample, omics_patient_mapping, prescriptions and labs.}

\item{index_range}{the number of days to look out from index date.}
}
\value{
A dataframe with the summary data and an excel file.
}
\description{
Flags SPARC biosamples based on date sample collected as either enrollment,
endoscopy, endoscopy and enrollment, 3 months after endoscopy, 3 months after
med change.
}
