% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_emr_subset.R
\name{sparc_emr}
\alias{sparc_emr}
\title{sparc_emr}
\usage{
sparc_emr(datadir, filename = "SPARC_EMR.xlsx", emr_codes = NULL)
}
\arguments{
\item{datadir}{directory where unzipped data is located.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{emr_codes}{the codes to search through in the extract}
}
\value{
A dataframe with the summary data and an excel file. If EMR codes are specified then a subset of those data are created for each domain where the code was found.
}
\description{
Subsets EMR data for specific EMR codes and provides spreadsheet with encounter or hist id
}
