% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_ends.R
\name{sparc_med_ends}
\alias{sparc_med_ends}
\title{sparc_med_ends}
\usage{
sparc_med_ends(medication)
}
\arguments{
\item{medication}{A dataframe with generated by sparc_med_filter}
}
\value{
A dataframe with the medication end date
}
\description{
#' Find medication discontinuation date from SPARC patient reported and electronic medical record data.
}
\details{
Medication start and stop dates are chosen independently from both
  eCRF and EMR sources. Medications with a start or stop date before 1980 are
  dropped.

  If eCRF data source OR EMR data source says that medication was stopped
  then count it as true. Consider medication end date for eCRF and EMR data
  sources and medication discontinue date for EMR.#'   The earliest of these
  dates is the end date. For EMR, if there is no prescription within 60 days
  or if there is a discontinuation date but a new prescription within 60 days
  then they are considered to be continuously on it.
}
