% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_filter.R
\name{sparc_med_filter}
\alias{sparc_med_filter}
\title{sparc_med_filter}
\usage{
sparc_med_filter(
  prescriptions,
  observations,
  demographics,
  encounter,
  med_groups = c("Aminosalicylates", "Antibiotics", "Antidiarrheals", "Biologic",
    "Corticosteroids", "Immunomodulators", "Other", "Probiotic",
    "Targeted synthetic small molecules")
)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}

\item{observations}{a dataframe with observation data usually generated using load_data.}

\item{demographics}{A dataframe with demographic data usually generated using load_data.}

\item{encounter}{a dataframe with encounter data usually generated using load_data.}

\item{med_groups}{A character string that filters the \code{\link{med_grp}} dataframe. Options are Aminosalicylates, Antibiotics, Antidiarrheals, Biologic,  Corticosteroids, Immunomodulators, Other, Probiotic. Default does not use any filtering.}
}
\value{
table with medications of interest from eCRF and EMR data
}
\description{
Logic to filter prescriptions table for medications of interest from the EMR and eCRF data sources.
}
\details{
Medication data is pulled forward if a patient answers "No" to the question "Are you currently taking any medication for your IBD?" on the quarterly survey.
The \code{\link{med_grp}} dataframe has the medication names to search for in the MEDICATION, OTHER_MEDICATION and SRC_DRUG_CODE_CONCEPT_NAME columns.
}
