% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_starts.R
\name{sparc_med_starts}
\alias{sparc_med_starts}
\title{sparc_med_starts}
\usage{
sparc_med_starts(medication, encounter)
}
\arguments{
\item{medication}{A dataframe with generated by sparc_med_filter}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}
}
\value{
A dataframe with the medication start date
}
\description{
Extract medication first instance of medication start dates from SPARC patient reported and electronic medical record data.
}
\details{
Medication start and stop dates are chosen independently from both
  eCRF and EMR sources. Medications with a start or stop date before 1980 are
  dropped. For EMR data, if a medication start date is missing, the visit
  encounter start date is used. These records are flagged in the column
  VISIT_ENCOUNTER_MED_START. If both data sources have information on a
  particular medication, the earliest start date is chosen. If the visit
  encounter start date is used for EMR and eCRF data is available, then the
  eCRF data source is used.
}
