% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_at_index.R
\name{sparc_medication}
\alias{sparc_medication}
\title{sparc_medication}
\usage{
sparc_medication(
  data,
  index_info = c("ENROLLMENT", "LATEST", "ENDOSCOPY", "OMICS", "BIOSAMPLE"),
  med_groups = c("Biologic", "Aminosalicylates", "Immunomodulators"),
  filename = "SPARC_MEDICATION.xlsx"
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, and prescriptions.}

\item{index_info}{A dataframe  with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Other options include, enrollment, latest, endoscopy, omics, biosample collection.}

\item{med_groups}{A list of medication groups of interest. For details, refer to \code{\link{sparc_med_filter}}.}

\item{filename}{the name of the output file. Must be .xlsx. If filename is NULL no output file will be written.}
}
\value{
A dataframe with medication predictions.
}
\description{
Finds the medications the participant is prescribed from the
electronic medical record and patient reported case report forms at a
specific (“index”) date.
}
\details{
When the index date is the date of consent, only the patient reported data is used.

If two medications of the same MOA overlap, the start of the 2nd medication is used as the end date of the previous medication.
If a filename is specified, an excel file will be written as a side effect.
}
