% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_at_enrollment.R
\name{sparc_medication_enrollment}
\alias{sparc_medication_enrollment}
\title{sparc_medication_enrollment}
\usage{
sparc_medication_enrollment(cohort, med)
}
\arguments{
\item{cohort}{dataframe with index_date}

\item{med}{dataframe generated by sparc_med_starts() function}
}
\value{
A dataframe and an excel file.
}
\description{
Finds the medications the participant is prescribed from the
electronic medical record and patient reported case report forms at enrollment.
}
