% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_scores_at_index_v4.R
\name{sparc_scores}
\alias{sparc_scores}
\title{sparc_scores}
\usage{
sparc_scores(
  datadir,
  index_info = "DATE_OF_CONSENT",
  filename = "SPARC_SCORES.xlsx",
  index_range = "30"
)
}
\arguments{
\item{datadir}{directory where unzipped data is located}

\item{index_info}{A dataframe with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Default is date of consent.}

\item{filename}{name of file to output. Must be xlsx format.}

\item{index_range}{numbers of days to look out of index date. For example, if you want data within +/- 30 days of the index date this value would be 30.}
}
\value{
A dataframe with scores at that date and an excel spreadsheet.
}
\description{
reads in SPARC data from IBD Plexus, and calculates diagnosis, phenotype, disease, and endoscopy scores at a specific (“index”) date.
}
