% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_smartform_summary.R
\name{sparc_sf}
\alias{sparc_sf}
\title{sparc_sf}
\usage{
sparc_sf(
  datadir,
  index_info = c("ALL", "ENROLLMENT", "LATEST", "ENDOSCOPY"),
  filename = "SMARTFORM_SUMMARY.xlsx"
)
}
\arguments{
\item{datadir}{Directory where unzipped data is saved.}

\item{index_info}{The time period of interest as either a default option or dataframe with two columns: DEIDENTIFIED_MASTER_PATIENT_ID, and index_date formatted as a date.
\itemize{
  \item ALL: All Smartform data. Each row will be a smartform entry.
  \item ENROLLMENT: Smartform data within 30 days of enrollment with one row per patient in the final table.
  \item LATEST: The latest Smartform data for each patient. If there are no changes at a patient visit, the Smartform is not updated and data is pulled forward from the previous visit.
  \item ENDOSCOPY: Smartform data within 30 days of endoscopy with one row per endoscopy visit in the final table.
}}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{index_range}{If index date is specified, the number of days to look out from index date.}
}
\value{
A dataframe and an excel file.
}
\description{
Summarizes the Smartform data at different time points for SPARC patients.
}
