% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_summary_at_index.R
\name{sparc_summary}
\alias{sparc_summary}
\title{sparc_summary}
\usage{
sparc_summary(
  datadir,
  index_info = "DATE_OF_CONSENT",
  filename = "SPARC_SUMMARY.xlsx",
  index_range = "30",
  emr_codes = NULL
)
}
\arguments{
\item{datadir}{directory where unzipped data is located.}

\item{index_info}{A dataframe  with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Default is date of consent.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{index_range}{the number of days to look out from index date.}

\item{emr_codes}{the codes to search through in the extract}
}
\value{
A dataframe with the summary data and an excel file. If EMR codes are specified then a subset of those data are created for each domain where the code was found.
}
\description{
Finds the medications the participant is prescribed from the electronic
medical record, patient reported case report forms and Smartform, lists other
Smartform data fields, calculates diagnosis, phenotype, disease, and
endoscopy scores and lists the encounter id or hist id at a specific
(“index”) date.
}
