% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steroid_use.R
\name{steroid_filter}
\alias{steroid_filter}
\title{For SPARC, this function further filters the steroid data to remove inhalation, nasal, eye drops and ointments.}
\usage{
steroid_filter(medication)
}
\arguments{
\item{medication}{A dataframe generated by sparc_med_filter()}
}
\value{
A filtered data frame like the one procuded by sparc_med_filter()
}
\description{
For SPARC, this function further filters the steroid data to remove inhalation, nasal, eye drops and ointments.
}
