% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steroid_use.R
\name{steroid_use}
\alias{steroid_use}
\title{steroid_use}
\usage{
steroid_use(prescriptions, observations, demographics, encounter, med)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}

\item{observations}{a dataframe with observations data usually generated using load_data.}

\item{demographics}{a dataframe with demographic data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}

\item{med}{A dataframe generated by joining the results from med_start() and med_end()}
}
\value{
A dataframe with the master patient id, data source, medication name, original dose and any doses reported over time for that source.
}
\description{
For SPARC, this function finds if a patient was on steroid at the time of
another medication in the medication journey. Steroid information must have
MED_START_DATE and MED_END_DATE or MED_DISCONT_START_DATE. Inhalation, eye drops and nasal
routes of administration removed.
}
