% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steroid_use.R
\name{steroid_use_at_index}
\alias{steroid_use_at_index}
\title{steroid_use_at_index}
\usage{
steroid_use_at_index(data, cohort)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, and prescriptions.}

\item{cohort}{A dataframe generated by sparc_medication}
}
\value{
A dataframe with the master patient id, data source, medication name, original dose and any doses reported over time for that source.
}
\description{
For SPARC, this function finds if a patient was on steroid at index date. Steroid information must have
MED_START_DATE and MED_END_DATE or MED_DISCONT_START_DATE. Inhalation, eye drops and nasal
routes of administration removed.
}
