% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{to_wide}
\alias{to_wide}
\title{to_wide}
\usage{
to_wide(df, y.var, value.var)
}
\arguments{
\item{df}{dataframe}

\item{y.var}{column names to be transposed}

\item{value.var}{values to be transposed}
}
\value{

}
\description{
Function to reshape table into a wide format. From rancho.
}
