% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_response.R
\name{tx_response}
\alias{tx_response}
\title{Create Treatment Response Table - all responses}
\usage{
tx_response(
  data,
  reports_path,
  all_events = T,
  save_path = "TX_RESPONSE.xlsx",
  time_category = F,
  month_range = 3,
  month_interest = 12,
  filtered_med = F,
  med_interest =
    "Adalimumab|Certolizumab Pegol|Etrasimod|Golimumab|Guselkumab|Infliximab|Mirikizumab|Natalizumab||Ozanimod|Risankizumab|Tofacitinib|Upadacitinib|Ustekinumab|Vedolizumab"
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, prescriptions and labs.}

\item{reports_path}{A path to already generated SPARC Reports.}

\item{all_events}{Defaults to T. Includes all disease activity measure events after a patient starts a medication.}

\item{save_path}{the name of the desired output file. Must be .xlsx.}

\item{time_category}{Defaults to F. Set to T when only disease activity measures from a specified time range are required.}

\item{month_range}{Month for time frame of interest, after medication start date. Defaults to 12 months.}

\item{month_interest}{Range of months around month of interest for which to include relevant disease activity measures. Defaults to 3 months.}

\item{filtered_med}{Defaults to F. Set to T to filter only for specific medications.}

\item{med_interest}{String of medications separated by | to search for when filtered_med = T. Possible medications are: Adalimumab, Azathioprine, Balsalazide, Certolizumab Pegol, Cyclosporine, Etrasimod, Golimumab, Guselkumab, Infliximab, Mercaptopurine, Mesalamine, Methotrexate, Mirikizumab, Natalizumab, Olsalazine, Ozanimod, Risankizumab, Sulfasalazine, Tacrolimus, Tofacitinib, Upadacitinib, Ustekinumab, Vedolizumab"}
}
\value{
Tx response data file
}
\description{
Create Treatment Response Table - all responses
}
\examples{
# upload data
# data <- load_data(datadir, cohort = "SPARC", domains = "ALL", data_type = "both")
# get treatment response values for patients on Tofacitinib and Upadacitinib 12 months after medication start date, with 3 month range
# jaki_12mo_txresponse <- tx_response(data, reports_path = "~/REPORTS/", all_events = F, save_path = "JAKI_12MO_TX_RESPONSE.xlsx", time_category = T, month_range = 3, month_interest = 12, filtered_med = T, med_interest = "Tofacitinib|Upadacitinib")
}
