% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcpdai.R
\name{wpcdai}
\alias{wpcdai}
\title{wpcdai}
\usage{
wpcdai(
  data,
  select.col = c("DIAG_CONCEPT_NAME", "DIAGNOSIS_DATE", "GENDER", "TYPE_OF_ENCOUNTER",
    "VISIT_ENCOUNTER_START_DATE", "AGE_AT_ENCOUNTER", "BIRTH_YEAR",
    "LAB_TEST_CONCEPT_NAME", "ASSAY NAME", "MEDICATION_NAME", "OBS_TEST_CONCEPT_NAME",
    "ANA_SITE_CONCEPT_NAME", "SRC_BIOSAMPLE_CONCEPT_NAME"),
  select.val = c("DIAG_STATUS_CONCEPT_NAME", "DIAGNOSIS_DATE", "GENDER",
    "TYPE_OF_ENCOUNTER", "VISIT_ENCOUNTER_START_DATE", "AGE_AT_ENCOUNTER", "BIRTH_YEAR",
    "TEST_RESULT_NUMERIC", "RAW DATA FILE NAME", "MED_START_DATE", "TEST_RESULT_NUMERIC",
    "TEST_RESULT_NUMERIC", "SAMPLE_STATUS"),
  filter.col = c("DIAG_CONCEPT_NAME", "OBS_TEST_CONCEPT_NAME"),
  filter.val = c("Disease Location", "Endoscopic Assessment - Deep Ulceration",
    "Endoscopic Assessment - Superficial Ulceration",
    "Endoscopic Assessment - Amount of Surface Ulcerated",
    "Endoscopic Assessment - Amount of Surface Involved", "Perianal Disease -", "EIM",
    "Disease Behavior - Stricturing/Fibrostenotic",
    "Disease Behavior - Internally Pentrating", "PCDAI")
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, prescriptions and labs.}
}
\value{
A dataframe with wPCDAI calcuation
}
\description{
Creates dataframe with patient and visit encounter ID's, wPCDAI calculation, and all PCDAI columns
}
\examples{
# upload data
# dat <- load_data(dir, cohort = "RISK", domains = c("ALL"), data_type = "BOTH")
# get dataframe with wpcdai
# wpcdai <- wpcdai(dat)


}
