% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_disease_scores_all_time.R
\name{calculate_disease_scores}
\alias{calculate_disease_scores}
\title{calculate_disease_scores}
\usage{
calculate_disease_scores(
  demographics,
  diagnosis,
  procedures,
  encounter,
  observations
)
}
\arguments{
\item{demographics}{demographics data}

\item{diagnosis}{diagnosis data from diagnosis_crf table}

\item{procedures}{procedures data from procedures_crf table}

\item{encounter}{encounter data from encounter_crf table}

\item{observations}{observations data from encounter_crf table}
}
\value{
An excel spreadsheet with the scores as different tabs and a list of dataframes with each score as a list element.
}
\description{
Creates an excel spreadsheet and list of data frames with diagnosis, sCDAI
(CD patients only), 6pt Mayo (UC Patients only), Manitoba, PGA, MES (UC
patients only), SES-CD (CD patients only), PRO2 (CD patients only) for SPARC
patients
Data can be loaded for this function using:
data <- load_data(datadir = datadir, cohort = "SPARC", domains = c("Demographics", "Diagnosis", "Procedures", "Encounter", "Observations"), data_type = "CRF")
}
