% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_calculations.R
\name{calculate_location_az}
\alias{calculate_location_az}
\title{calculate_location_az}
\usage{
calculate_location_az(observations, encounter, window = 90)
}
\arguments{
\item{observations}{observations table usually generated using load_data}

\item{encounter}{encounter table usually generated using load_data}

\item{window}{Time window (in days) in which observations labeled "unknown" can be carried over.}
}
\value{
A dataframe with all CD locations could be calculated regardless of IBD diagnosis.
}
\description{
Calculates disease locations from SPARC data with criteria specific to AZ
}
\details{
Contradictions and ambiguities in a location's involvement are resolved as follows:
If more than one record for a location exists for the same patient on the
same date, any record of a "yes" is taken as evidence for this location
showing disease activity. In absence of any "yes", any record of a "no" is
taken of evidence of absence of evidence. Otherwise, any occurrence of
"unknown" among the records is collapsed to a single "unknown". Any remaining
combination of values is considered "INVALID".

The window variable can be used to leverage the relative stability of the
involvement of locations in Crohn's disease. Missing (i.e. "Unknown") values
will be replaced if other observations exist in the specified time window.
Observations in the future of the visit with missing records are given
precedence over values in the past.

The columns colonic involvement and upper GI are constructed assuming that
explicit records of the absence of involvement of a location may have been
omitted. Any record of an "no" in relevant sub-locations are considered
sufficient to assume "unknown" in related locations imply "no" as long as a
"yes" is located anywhere else. For instance, if records indicate "no" for
left colonic involvement, and all other records for colonic sites are unknown.
This "no" will be considered sufficient to assume no colonic involvement if
involvement of the ileum or upper GI tract is recorded.
}
