% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_of_interest.R
\name{extract_latest}
\alias{extract_latest}
\title{extract_latest}
\usage{
extract_latest(encounter)
}
\arguments{
\item{encounter}{A dataframe with encounter data.}
}
\value{
A dataframe with  date and type of most recent encounter.
}
\description{
Extract most recent smartform or eCRF date for SPARC.
}
