% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyibdhist_emr_extract.R
\name{familyibdhist_emr_extract}
\alias{familyibdhist_emr_extract}
\title{familyibdhist_emr_extract}
\usage{
familyibdhist_emr_extract(data, relation = NULL)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, and the Patient History tables. Only EMR data needs to be uploaded.}

\item{relation}{Define whether to search for "FIRST_DEG", only including first degree family members.The default is to search for all family members.}
}
\value{
A list of dataframes with all relevant EMR Patient History records.
}
\description{
A function to extract records from the Family History EMR table indicating a family member with a history of IBD. User can define to include all family members or only First Degree Family members.
First degree is defined as: Biological Parent, Sibling or Child.
}
\examples{

# example for patient family history

# famhist = familyibdhist_emr_extract(data)
#example for only first degree family member history

#famhist1stdeg = familyibdhist_emr_extract(data,relation="FIRST_DEG")
}
