% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plexus_data_loading_script.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(
  datadir,
  cohort = c("RISK", "QORUS", "SPARC"),
  domains = c("ALL"),
  data_type = c("BOTH", "CRF", "EMR")
)
}
\arguments{
\item{datadir}{The directory where data is saved.}

\item{cohort}{The cohort to load. Either RISK, QORUS, or SPARC.}

\item{domains}{The domains to load. Default is "ALL". Must be a character string. Other valid values are patterns matching input file names common examples are:  "demographics", "diagnosis", "encounter", "procedures", "observations", "biosample", "omics_patient_mapping", "prescriptions".}

\item{data_type}{The data source to load either case report forms (including covid survey), electronic medical record, or both. Options are both, crf or emr.}
}
\value{
A list of dataframes for each domain. If both sources are loaded, emr and crf data are combined.
}
\description{
Load unzipped DDM txt or csv files.If multiple cohorts are unzipped in one directory from different times, only the most recent one will load.
}
