% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_grp}
\alias{med_grp}
\title{med_grp}
\format{
A data frame with 105 rows and 3 variables:
\describe{
  \item{MEDICATION_NAME}{Medication Name as it appears in SF, ECRF or EMR}
  \item{med_type}{The class of medication}
  \item{new_med_name}{Standardized medication name}

}
}
\usage{
med_grp
}
\description{
A dataframe containing the coding for Medications. Allows to group the same
medications together if they have different names (generic v. scientific)
from different source. Infliximab Biosimilars added March 2022.
}
\keyword{datasets}
