% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qorus_scores.R
\name{qorus_scores}
\alias{qorus_scores}
\title{Create QORUS Scores Table - all responses}
\usage{
qorus_scores(
  data,
  filename = "QORUS_SCORES.xlsx",
  location_logic = "CCF",
  export = T
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, and observations. Only CRF data needs to be uploaded.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{location_logic}{specify which CD location function to use. If using logic developed by AZ then use location_logic = "AZ".}

\item{export}{if TRUE then writes the data frame to its own excel file}
}
\value{
qorus scores table
}
\description{
Create QORUS Scores Table - all responses
}
