% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qorus_summary_at_index.R
\name{qorus_summary}
\alias{qorus_summary}
\title{qorus_summary}
\usage{
qorus_summary(
  datadir,
  index_info = "DATE_OF_CONSENT",
  filename = "QORUS_SUMMARY.xlsx",
  index_range = "30",
  emr_codes = NULL
)
}
\arguments{
\item{datadir}{Directory where unzipped data is saved.}

\item{index_info}{A data frame with DEIDENTIFIED_MASTER_PATIENT_ID and a variable named index_date. Default is the Date of Consent.}

\item{filename}{The name of the output file.}

\item{index_range}{The number of days to look from index date. Default is 30 days.}

\item{emr_codes}{EMR codes of interest as a character string. Can be an ICD10, LOINC or CPT code.}
}
\value{
Returns a dataframe and excel spreadsheet. If EMR codes are specified, subset of data with that code.
}
\description{
Summarizes QORUS data at a specific date.
}
