% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_helper_functions.R
\name{risk_meds_at_visit}
\alias{risk_meds_at_visit}
\title{risk_medication_at_visit}
\usage{
risk_meds_at_visit(data)
}
\arguments{
\item{data}{RISK data loaded using the ibdplexus::load_data function with all domains}
}
\value{
A dataframe of RISK patients with a column MEDICATIONS_AT_VISIT, deidentified patient and encounter ID's, and date columns
}
\description{
create table with column for RISK patients that lists the medications they were on at a visit
}
