% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_starts.R
\name{sparc_med_starts}
\alias{sparc_med_starts}
\title{sparc_med_starts}
\usage{
sparc_med_starts(
  prescriptions,
  demographics,
  observations,
  encounter,
  export = FALSE
)
}
\arguments{
\item{prescriptions}{A dataframe with prescriptions data usually generated using load_data.}

\item{demographics}{a dataframe with demographic data usually generated using load_data.}

\item{observations}{a dataframe with observations data usually generated using load_data.}

\item{encounter}{a dataframe with encounter data usually generated using load_data.}

\item{export}{if excel spreadsheet should be exported. FALSE is default.}
}
\value{
A dataframe with the first medication start date for each drug.
}
\description{
Extract medication first instance of medication start dates from SPARC patient reported and electronic medical record data. This
is only for long term medications including Biologics, Aminosalicylates, and
Immunomodulators.
}
\details{
Medication start and stop dates are chosen independtly from both eCRF and EMR sources. Medications with a start or stop date before 1980 are dropped.
For EMR data, if a medication start date is missing, the visit encounter start date is used. These records are flagged in the column VISIT_ENCOUNTER_MED_START.

Bionaive is 1 if a patient has no prior reported biologics (including JAK inhibitors and Tofacitinib). Started after enrollment is 1 if the medication start date is after the date of consent.

If a patient has medication information for the same drug from eCRF and EMR, the eCRF data is preferred and used to generate MED_START_DATE and MED_END_DATE. If only EMR data is available for that medication, then EMR data is used.
Any overlap between medications is reported along with the number of days the medications overlap.
}
