% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_medication_at_index.R
\name{sparc_medication}
\alias{sparc_medication}
\title{sparc_medication}
\usage{
sparc_medication(
  datadir,
  index_info = "DATE_OF_CONSENT",
  filename = "SPARC_MEDICATION.xlsx",
  index_range = "30"
)
}
\arguments{
\item{datadir}{directory where unzipped data is saved.}

\item{index_info}{A dataframe with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Default is date of consent.}

\item{filename}{the name of the output file.}

\item{index_range}{the number of days to look out from index date.}
}
\value{
A dataframe with medication predictions and an excel file.
}
\description{
Reads in SPARC data from IBD Plexus, and finds the medications the participant is prescribed from the
electronic medical record and patient reported case report forms at a
specific (“index”) date.
}
\details{
When the index date is the date of consent, only the patient reported data is used.
}
