% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_scores_at_index.R
\name{sparc_scores}
\alias{sparc_scores}
\title{sparc_scores}
\usage{
sparc_scores(
  data,
  index_info = c("ENROLLMENT", "LATEST", "ENDOSCOPY", "OMICS", "BIOSAMPLE"),
  filename = "SPARC_SCORES.xlsx",
  index_range = "30",
  location_logic = "CCF",
  export = T
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, and observations. Only CRF data needs to be uploaded.}

\item{index_info}{A dataframe  with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Other options include, enrollment, latest, endoscopy, omics, biosample collection.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{index_range}{the number of days to look out from index date. If multiple of the same score fall within this range then the one closest to the index date is picked. Default is 30 days.}

\item{location_logic}{specify which CD location function to use. If using logic developed by AZ then use location_logic = "AZ".}

\item{export}{if TRUE then writes the data frame to its own excel file}
}
\value{
A dataframe with scores at that date and an excel spreadsheet.
}
\description{
Summarize the disease severity and endoscopy scores closest to a specific index date within a specific range of time.
}
\details{
If multiple of the same score type are reported on the same date from the same data source, the more severe score is taken. If multiple scores are reported from two different sources, the Smartform data source is preferred over the eCRF data source.

If endoscopy, omics or biosample collection date are chosen as index date then endoscopy, omics_patient_mapping or biosample tables must be in the data parameter.
}
