% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparc_summary_at_index.R
\name{sparc_summary}
\alias{sparc_summary}
\title{sparc_summary}
\usage{
sparc_summary(
  data,
  index_info = c("ENROLLMENT", "LATEST", "ENDOSCOPY", "OMICS", "BIOSAMPLE"),
  filename = "SPARC_SUMMARY.xlsx",
  index_range = "30"
)
}
\arguments{
\item{data}{A list of dataframes, typically generated by a call to load_data. Must include demographics, diagnosis, encounter, procedures, observations, biosample, omics_patient_mapping, and prescriptions.}

\item{index_info}{A dataframe  with DEIDENTIFIED_MASTER_PATIENT_ID and a variable index_date.  Other options include, enrollment, latest, endoscopy, omics, biosample collection.}

\item{filename}{the name of the output file. Must be .xlsx.}

\item{index_range}{the number of days to look out from index date.}
}
\value{
A dataframe with the summary data and an excel file.
}
\description{
Finds the medications the participant has been exposed to from the Smartform, lists other
Smartform, EMR and eCRF data fields at a particular point in time
}
\examples{
# upload data
data <- load_data(datadir, cohort = "SPARC", domains = c("demographics", "diagnosis", "encounter", "procedures", "observations", "biosample", "omics_patient_mapping", "prescriptions"), data_type = "both")
# get summary data within 30 days of endoscopy
endoscopy_summary <- sparc_summary(data = data, index_info = "ENDOSCOPY")
}
