% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcpdai.R
\name{wpcdai}
\alias{wpcdai}
\title{wpcdai}
\usage{
wpcdai(observations, encounter, labs)
}
\arguments{
\item{observations}{A dataframe with observation data usually generated using load_data.}

\item{encounter}{A dataframe with encounter data usually generated using load_data.}

\item{labs}{A dataframe with lab data usually generated using load_data.}
}
\value{
A dataframe with wPCDAI calcuation
}
\description{
Creates dataframe with patient and visit encounter ID's, wPCDAI calculation, and all PCDAI columns
}
\examples{
# upload data
# dat <- load_data(dir, cohort = "RISK", domains = c("ALL"), data_type = "BOTH")
# observations <- dat$observations
# encounter <- dat$encounter
# labs <- dat$labs
# get dataframe with wpcdai
# wpcdai <- wpcdai(observations, encounter, labs)


}
